Ajustes Global
eu retirei o global styles de um .js extra e coloquei no layout onde era requisitado
