+++ package.json
@@ -5,15 +5,16 @@
   "version": "0.1.0",
   "author": "Kyle Mathews <mathews.kyle@gmail.com>",
   "dependencies": {
+    "babel-plugin-styled-components": "^1.10.7",
     "gatsby": "^2.21.37",
     "gatsby-image": "^2.4.4",
-    "gatsby-plugin-manifest": "^2.4.5",
+    "gatsby-plugin-manifest": "^2.4.8",
     "gatsby-plugin-offline": "^3.2.3",
-    "gatsby-plugin-react-helmet": "^3.3.1",
-    "gatsby-plugin-sharp": "^2.6.4",
-    "gatsby-plugin-styled-components": "^3.3.1",
-    "gatsby-source-filesystem": "^2.3.4",
-    "gatsby-transformer-sharp": "^2.5.2",
+    "gatsby-plugin-react-helmet": "^3.3.2",
+    "gatsby-plugin-sharp": "^2.6.8",
+    "gatsby-plugin-styled-components": "^3.3.2",
+    "gatsby-source-filesystem": "^2.3.7",
+    "gatsby-transformer-sharp": "^2.5.3",
     "prop-types": "^15.7.2",
     "react": "^16.12.0",
     "react-dom": "^16.12.0",
